+++ app/src/main/java/com/pinit/api/PinItAPI.java
@@ -4,6 +4,9 @@ import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.pinit.api.errors.APIError;
 import com.pinit.api.errors.BadRequestError;
+import com.pinit.api.listeners.LoginListener;
+import com.pinit.api.listeners.NetworkListener;
+import com.pinit.api.listeners.PinLikedByMeListener;
 import com.pinit.api.models.Comment;
 import com.pinit.api.models.Like;
 import com.pinit.api.models.Pin;
