+++ app/src/main/java/com/pinit/api/listeners/LoginListener.java
@@ -1,4 +1,4 @@
-package com.pinit.api;
+package com.pinit.api.listeners;
 
 import com.pinit.api.errors.APIError;
