+++ app/src/main/java/com/pinit/api/listeners/NetworkListener.java
@@ -1,4 +1,4 @@
-package com.pinit.api;
+package com.pinit.api.listeners;
 
 import com.pinit.api.errors.APIError;
