+++ app/src/main/java/com/pinit/api/listeners/PinLikedByMeListener.java
@@ -1,4 +1,4 @@
-package com.pinit.api;
+package com.pinit.api.listeners;
 
 public interface PinLikedByMeListener {
     void isLikedByMe();
