+++ app/src/main/java/com/pinit/pinitmobileapp/MapsActivity.java
@@ -14,13 +14,11 @@ import android.location.Location;
 import android.location.LocationManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.design.widget.CheckableImageButton;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.FragmentActivity;
 import android.support.v7.widget.SwitchCompat;
 import android.support.v7.widget.AppCompatButton;
-import android.text.TextUtils;
 import android.view.View;
 import android.widget.*;
 
@@ -32,7 +30,7 @@ import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.*;
 import com.google.maps.android.heatmaps.HeatmapTileProvider;
 import com.pinit.api.errors.APIError;
-import com.pinit.api.NetworkListener;
+import com.pinit.api.listeners.NetworkListener;
 import com.pinit.api.PinItAPI;
 import com.pinit.api.models.Comment;
 import com.pinit.api.models.Pin;
